UPDATED STRUCTURE
This commit fixes a major bug that prevented the GUI from being constructed. Being the silly man that I am, I had the abstract panel inherit the abstract screen because there was so much shared code! But this unfortunately made Swing
quite confused, and got it stuck in infinite loops that caused StackOverflow errors. So, I separated the panel and the screen and may one day try to find a different way to bring them together--inheritance was not that way, as that
one video by Code Aesthetics I watched should have taught me (although this was less an issue with coupling I think, more an issue with self-reference).

Anyway, now the GUI actually does show up, so this is the true good first commit where the tester can properly be used.