UPDATED MANY THINGS
This is still the setup stage, but this commit updates various things, fixes bugs, and allows the actual displaying of slow-printed text. The next step is going to be setting up the input system, which I think I will be able to do
based on what I have architected so far.

Basically, there will be a Listener interface that is implemented by individuated listener classes that will handle the type of input. Then the proper listener class will be added to the abstract objects as a reference, and that will
then be the way that certain things are done; i.e. abstractPanel will have the listener, and will always take in a InputEvent Event since those are its children, whereas Screen will always take in PanelEvent event because those are
always its children, and window will always take in ScreenEvent event.

I will need to create an Event enum or set of classes, but that's about it.